In fx2usbdfu and fx2usbmassstor, turn off overlay for setup handler.
The setup handlers are leaf functions that are inferred to be non-
reentrant, but this isn't actually safe. The memory for parameters
(other than first) of non-reentrant functions is allocated in the OVR
memory area, which means that it can be shared with other functions
inferred as non-reentrant.

This causes nasty race conditions that tend to crash the device.