Migrate from setup.py to pyproject.toml.
This commit also vendors `boot-cypress.ihex`, built with sdcc 4.0.0
from the current firmware/boot-cypress/ source. This is normally not
something I would do, but packaging it as a part of the Python build
process has always been unreliable, and it's essentially static, so
it should be fine to do this.